Pre-Meeting Update
-Imports for Camera

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Aut_Encoder.java
@@ -46,6 +46,21 @@ import com.qualcomm.robotcore.util.Range;
 import android.app.Activity;
 import android.view.View;
 
+import org.firstinspires.ftc.robotcontroller.external.samples.ConceptVuforiaNavigation;
+import org.firstinspires.ftc.robotcore.external.ClassFactory;
+import org.firstinspires.ftc.robotcore.external.matrices.OpenGLMatrix;
+import org.firstinspires.ftc.robotcore.external.matrices.VectorF;
+import org.firstinspires.ftc.robotcore.external.navigation.AngleUnit;
+import org.firstinspires.ftc.robotcore.external.navigation.AxesOrder;
+import org.firstinspires.ftc.robotcore.external.navigation.AxesReference;
+import org.firstinspires.ftc.robotcore.external.navigation.Orientation;
+import org.firstinspires.ftc.robotcore.external.navigation.RelicRecoveryVuMark;
+import org.firstinspires.ftc.robotcore.external.navigation.VuMarkInstanceId;
+import org.firstinspires.ftc.robotcore.external.navigation.VuforiaLocalizer;
+import org.firstinspires.ftc.robotcore.external.navigation.VuforiaTrackable;
+import org.firstinspires.ftc.robotcore.external.navigation.VuforiaTrackableDefaultListener;
+import org.firstinspires.ftc.robotcore.external.navigation.VuforiaTrackables;
+
 /**
  * This file illustrates the concept of driving a path based on encoder counts.
  * It uses the common Pushbot hardware class to define the drive on the robot.
@@ -271,7 +286,7 @@ public class Aut_Encoder extends LinearOpMode {
 
         //assuming that the servo is on the right side of the robot
 //        sense();
-        int timeToPicto = 0;
+        double timeToPicto = 0;
         if(isBlue == true) { //senses backwards
             if(ballSensor.blue()-50 > ballSensor.red()) { //if is blue
                 driveForward(50,1);
